+++ setup.py
@@ -60,7 +60,7 @@ setup(name='pyFF',
       url='http://blogs.mnt.se',
       license='BSD',
       setup_requires=['nose>=1.0'],
-      tests_require=['pbr', 'coverage', 'nose>=1.0', 'mock', 'mako', 'mockredispy', 'testfixtures'],
+      tests_require=['pbr', 'coverage', 'nose>=1.0', 'mock', 'mako', 'mockredispy', 'testfixtures', 'wsgi_intercept'],
       test_suite="nose.collector",
       packages=find_packages('src'),
       package_dir={'': 'src'},
